Removed a number of tests since they are useless.
Now I've brushed up on some probability theory and according to my old
math books we're dealing with the hypergeometric distribution and this
problem is basically the same as one of the examples in the book so I
expect this to be quite straight forward.

First of all we need a function to calculate how many ways we can
draw Y samples from a set of X in total.