Finsihed the implementation of solve.
A lot of stuff has happened since last check in. A bit too much really.
I renamed c to comb and rewrote the implementation. It turned out to be
a quite bad idea to use itertools.combinations to calculate the number
of possible combinations. It took too long once the numbers starting
getting close to 100 and it can quite easily be calculated anyway.

I also renamed p to prob.

What's left now is to write the stdin functionality, comment out all
doctest calls and perhaps a little refactoring.