Added implementation to match the failing test.
I also realized that my first failing test where I expected the result
to be [2011, 2012
] was a bit strange. [1009, 1017] was an equally good answer. Now, after
having finished the implementation, I've modified the test to match
the result. I'm still not sure I can be 100% certain on how the dict
implementation will behave with regards to sorting 'equal' entries so
this test might have to be re-written.